Update personal_finance.c
1. Created inputTransactionDetails in order to solve the Code Duplication: 

The logic for adding income and expense transactions is duplicated. This can be refactored to reduce the repetition. I added structure inputTransactionDetails and let users to input their transaction. Then there is no need to repeat the same code in main function for case1 and case2

2.  I adjust the function signature of createTransaction to accept const char* for the type and category arguments since you're trying to pass a const char* argument from inputTransactionDetails to createTransaction.

3. I adjust the function viewTransactionsAmout data type since category is a string (or a character array), and should be using the %s format specifier to print it.

4. I created isValidDate function to check if the date is valid after the user input it.

5. I've created updateTransaction function to update each transaction user had.

6.I have created case 8 let user to input their info also modify with isValidDate function to check if the date is valid or not.

especially I've added more explanatory text to guide the user through the updating process.